+++ inc/parsing.h
@@ -6,7 +6,7 @@
 /*   By: mrollo <mrollo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/15 14:46:05 by anarebelo         #+#    #+#             */
-/*   Updated: 2023/05/03 19:20:50 by mrollo           ###   ########.fr       */
+/*   Updated: 2023/05/05 13:36:57 by mrollo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,15 +19,21 @@
 
 void	check_args(int argc, char **argv, t_master *master);
 int		parse(char *path, t_map *map);
-int     read_file(char *path, t_map *map);
+int		read_file(char *path, t_map *map);
 char	**create_map(int x, int y);
 int		fill_map(t_map *map);
 int		check_map(char **mtx, t_map *map);
+int		check_nb_rows(char *row, int cols);
+int		check_nb_cols(char **map, int rows, int j);
 char	*tex_parse(char *str);
-// int		*color_arr(char *line);
-int	    *parse_color_array(char *str);
+int		*parse_color_array(char *str);
 int		check_color(int *arr);
-void	free_tab(char** tab);
+void	free_tab(char **tab);
 void	error_control(char *msj);
+void	save_ini_pos(t_map *map);
+int		ft_isspace(char *line);
+int		check_path(char *path);
+int		color_parse(char *line, char a, t_map *map);
+char	*clean_color(char *str);
 
 #endif
