+++ srcs/parsing/check_rows_cols.c
@@ -1,0 +1,44 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   check_rows_cols.c                                  :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: mrollo <mrollo@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/05/05 13:21:59 by mrollo            #+#    #+#             */
+/*   Updated: 2023/05/05 13:24:00 by mrollo           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3D.h"
+#include "parsing.h"
+
+int	check_nb_rows(char *row, int cols)
+{
+	int	i;
+
+	i = 0;
+	while (i < cols)
+	{
+		if (row[i] == ' ' || row[i] == '1')
+			i++;
+		else
+			return (1);
+	}
+	return (0);
+}
+
+int	check_nb_cols(char **map, int rows, int j)
+{
+	int	i;
+
+	i = 0;
+	while (i < rows)
+	{
+		if (map[i][j] == ' ' || map[i][j] == '1')
+			i++;
+		else
+			return (1);
+	}
+	return (0);
+}
