+++ srcs/parsing/map_check.c
@@ -6,42 +6,12 @@
 /*   By: mrollo <mrollo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/26 15:24:56 by mrollo            #+#    #+#             */
-/*   Updated: 2023/05/03 18:18:58 by mrollo           ###   ########.fr       */
+/*   Updated: 2023/05/05 13:23:01 by mrollo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3D.h"
 #include "parsing.h"
-
-int	check_nb_rows(char *row, int cols)
-{
-	int	i;
-
-	i = 0;
-	while (i < cols)
-	{
-		if (row[i] == ' ' || row[i] == '1')
-			i++;
-		else
-			return (1);
-	}
-	return (0);
-}
-
-int	check_nb_cols(char **map, int rows, int j)
-{
-	int	i;
-
-	i = 0;
-	while (i < rows)
-	{
-		if (map[i][j] == ' ' || map[i][j] == '1')
-			i++;
-		else
-			return (1);
-	}
-	return (0);
-}
 
 int	check_round_2(char **map, int x, int y)
 {
